requirements.txt
# Core Web Framework (Lightweight)
flask==2.3.3
flask-cors==3.0.10
gunicorn==21.2.0

# Database (Essential only)
pymongo==4.6.1

# AWS (Minimal)
boto3==1.34.81
botocore==1.34.81

# OpenAI and LangChain (Core only)
openai==1.12.0
langchain>=0.1.0
langchain-community>=0.0.27
langchain-core>=0.1.0
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.1

# PDF Generation (Lightweight)
reportlab==4.0.0

# HTTP Client (Essential)
httpx==0.27.0

# Environment and Utils (Minimal)
python-dotenv==1.0.0
pytz==2024.1

# PDF processing (One library is enough)
pypdf==3.17.1

# Vector store (Lightweight)
faiss-cpu==1.7.4

# Utilities (Essential only)
pydantic==2.6.1
tiktoken==0.6.0

# Core ML (Minimal)
numpy==1.24.4
scikit-learn==1.3.0

psutil==6.0.0




# Remove heavy packages:
# - Removed pandas (use numpy instead)
# - Removed sentence-transformers (use simpler text processing)
# - Removed torch (use CPU-only alternatives)
# - Removed transformers (use simpler text processing)
# - Removed fastapi (using flask only)
# - Removed unnecessary dependencies


